[#34] implement packet restructuring
